issue 23 fixed getLength() to return int for

+++ trunk/dsm2-input-model/src/gov/ca/dsm2/input/model/Channel.java
@@ -38,7 +38,7 @@ public class Channel implements Serializable {
 		return id;
 	}
 
-	public double getLength() {
+	public int getLength() {
 		return length;
 	}
